fix(ui): make /accounts/[ref] static-export compatible + export-build CI gate
The release "Build dashboard" step builds the UI with
`NEXT_EXPORT=1 npm run build` (output: export) for the Python package.
`/accounts/[ref]` — the app's only dynamic route — failed that build:

    Error: Page "/accounts/[ref]" is missing "generateStaticParams()"
    so it cannot be used with "output: export" config.

Standalone PR CI (`npm run build`) does not enforce generateStaticParams,
so the route passed PR CI and only broke at release time.

A `"use client"` page cannot export `generateStaticParams`, so split the
route:

- AccountDetailClient.tsx keeps the client body (reads the ref via
  useParams at runtime).
- page.tsx becomes a server component that supplies generateStaticParams
  and renders the client child. Next 16 rejects an empty param list under
  output: export, so it emits one inert placeholder segment; real refs are
  read client-side from the URL. dynamicParams is false (output: export
  forbids true).

Prevent recurrence: add a "UI static-export build (release parity)" step
to the UI Validate job that runs `NEXT_EXPORT=1 npm run build`, so any
future export-incompatible route is caught in PR CI instead of at release.
It runs last because the export build overwrites .next/out that the
bundle-budget and E2E steps consume.

Verified: NEXT_EXPORT=1 build green (route SSGs /accounts/_), standalone
build green (.next/standalone present), vitest 623 passed, eslint clean.
